Implement mode management functionality and refactor STOToolsKeybindManager
- Introduced a new `modeManagement.js` module to encapsulate mode switching and UI updates based on the current environment, enhancing code organization and maintainability.
- Refactored `STOToolsKeybindManager` in `app.js` to integrate the new mode management functionality, improving separation of concerns within the application.
- Removed redundant mode-related methods from `STOToolsKeybindManager`, streamlining the code and enhancing readability.
- These changes significantly improve the application's structure and user experience by providing a dedicated module for mode management.
